Create ResultsChart.svelte
