+++ frontend/src/routes/rooms/[roomId]/ResultsChart.svelte
@@ -1,0 +1,29 @@
+<script lang="ts">
+  import Chart from "$lib/components/Chart.svelte";
+  import type { Room } from "$lib/services/rooms";
+
+  export let roomData: Room;
+
+  let chartLabels: string[] = [];
+  let chartDataSeries: number[] = [];
+
+  $: {
+    chartLabels = roomData?.validSizes ?? [];
+    const currentVotes =
+      roomData?.users.filter((user) => user.hasVote).map((user) => user.vote) ??
+      [];
+    const valueFrequencies = currentVotes?.reduce((map, vote) => {
+      map.set(vote, (map.get(vote) ?? 0) + 1);
+      return map;
+    }, new Map<string, number>());
+    chartDataSeries = chartLabels.map(
+      (vote) => valueFrequencies.get(vote) ?? 0
+    );
+  }
+</script>
+
+<Chart
+  labels={chartLabels}
+  series={chartDataSeries}
+  options={{ distributeSeries: true }}
+/>
